requirements.txt
pulumi>=3.0.0,<4.0.0
proxmoxer-2.2.0
crypt-4.3.0 
cffi-1.17.1 
cryptography-45.0.6 
invoke-2.2.0 
paramiko-4.0.0 
pycparser-2.22 
pynacl-1.5.0
arpeggio-2.0.2 
attrs-25.3.0 
parver-0.5 
pulumi_proxmoxve-7.0.0
mariadb-1.1.13 
packaging-25.0
pycryptodome-3.23.0